Fixed up remaining problems with Arena with Hill. Player taunts no longer cut off. Enemies have minimap Icons. Weapon damage and enemy damage have been adjusted to make for a better fight.  player speed adjusted.  The sound was getting cut off because if you do "Audio.play()" only one thing can play at a time. "Audio.PlayOneShot()" allows for multiple sound events. ALSO the player now has two Audio Sources ON PURPOSE. This allows for you to play walking sound then stop it when you go idle WITHOUT stopping the sounds of a taunt or attack. PHEW

+++ Unity Project Files/Assets/!Scripts/PlayerHUD.js
@@ -88,7 +88,7 @@ function OnGUI()
 	GUI.DrawTexture(new Rect((Screen.width/2)-88, Screen.height-82, 176, 20),healthBarBackground
 	,ScaleMode.StretchToFill);
 	//Health Bar Red Color
-	var healthBarWidth : int = 176 * ((playerScript.currentHealth*1.0) / playerScript.maximumHealth);
+	var healthBarWidth : double = 176 * ((playerScript.currentHealth*1.0) / playerScript.maximumHealth);
 	GUI.DrawTexture(new Rect((Screen.width/2)-88, Screen.height-82, healthBarWidth, 20),canAttackEnemyColor
 	,ScaleMode.StretchToFill);
 	//Health Bar Text
